refactor(validation-schema): organize schema definitions
