+++ .github/workflows/test.yml
@@ -5,7 +5,6 @@ on:
     paths-ignore:
     - data-migrations/**
     - helm/**
-    - localdev/**
     - terraform/**
     - tools/**
