+++ app/app/api/private-cloud/products/_operations/create.ts
@@ -1,8 +1,8 @@
 import { Session } from 'next-auth';
 import { OkResponse, UnauthorizedResponse } from '@/core/responses';
 import createRequest from '@/request-actions/private-cloud/create-request';
-import { PrivateCloudCreateRequestBody } from '@/schema';
 import { sendCreateRequestEmails } from '@/services/ches/private-cloud/email-handler';
+import { PrivateCloudCreateRequestBody } from '@/validation-schemas/private-cloud';
 
 export default async function createOp({ session, body }: { session: Session; body: PrivateCloudCreateRequestBody }) {
   const { user, permissions, ministries } = session;
