+++ app/app/api/private-cloud/products/_operations/update.ts
@@ -5,9 +5,9 @@ import { BadRequestResponse, OkResponse, UnauthorizedResponse } from '@/core/res
 import { sendRequestNatsMessage } from '@/helpers/nats-message';
 import { getPrivateCloudProduct } from '@/queries/private-cloud-products';
 import editRequest from '@/request-actions/private-cloud/edit-request';
-import { PrivateCloudEditRequestBody } from '@/schema';
 import { sendEditRequestEmails, sendRequestApprovalEmails } from '@/services/ches/private-cloud/email-handler';
 import { subscribeUsersToMautic } from '@/services/mautic';
+import { PrivateCloudEditRequestBody } from '@/validation-schemas/private-cloud';
 import { putPathParamSchema } from '../[licencePlate]/schema';
 
 export default async function updateOp({
