+++ app/app/api/private-cloud/requests/[id]/decision/route.test.ts
@@ -2,7 +2,6 @@ import { expect } from '@jest/globals';
 import { $Enums } from '@prisma/client';
 import { createSamplePrivateCloudProductData } from '@/helpers/mock-resources';
 import { pickProductData } from '@/helpers/product';
-import { QuotaCpuEnum, QuotaMemoryEnum, QuotaStorageEnum } from '@/schema';
 import { mockSessionByEmail, mockSessionByRole } from '@/services/api-test/core';
 import { provisionPrivateCloudProject } from '@/services/api-test/private-cloud';
 import {
@@ -11,6 +10,7 @@ import {
   deletePrivateCloudProject,
 } from '@/services/api-test/private-cloud/products';
 import { makePrivateCloudRequestDecision } from '@/services/api-test/private-cloud/requests';
+import { CpuQuota, MemoryQuota, StorageQuota } from '@/validation-schemas/private-cloud';
 
 const fieldsToCompare = [
   'name',
@@ -28,15 +28,15 @@ const fieldsToCompare = [
 ];
 
 const oldDevelopmentQuota = {
-  cpu: QuotaCpuEnum.enum.CPU_REQUEST_1_LIMIT_2,
-  memory: QuotaMemoryEnum.enum.MEMORY_REQUEST_4_LIMIT_8,
-  storage: QuotaStorageEnum.enum.STORAGE_2,
+  cpu: CpuQuota.enum.CPU_REQUEST_1_LIMIT_2,
+  memory: MemoryQuota.enum.MEMORY_REQUEST_4_LIMIT_8,
+  storage: StorageQuota.enum.STORAGE_2,
 };
 
 const newDevelopmentQuota = {
-  cpu: QuotaCpuEnum.enum.CPU_REQUEST_1_LIMIT_2,
-  memory: QuotaMemoryEnum.enum.MEMORY_REQUEST_4_LIMIT_8,
-  storage: QuotaStorageEnum.enum.STORAGE_2,
+  cpu: CpuQuota.enum.CPU_REQUEST_1_LIMIT_2,
+  memory: MemoryQuota.enum.MEMORY_REQUEST_4_LIMIT_8,
+  storage: StorageQuota.enum.STORAGE_2,
 };
 
 const productData = {
