+++ app/app/api/public-cloud/products/_operations/create.ts
@@ -1,8 +1,8 @@
 import { Session } from 'next-auth';
 import { OkResponse, UnauthorizedResponse } from '@/core/responses';
 import createRequest from '@/request-actions/public-cloud/create-request';
-import { PublicCloudCreateRequestBody } from '@/schema';
 import { sendCreateRequestEmails } from '@/services/ches/public-cloud/email-handler';
+import { PublicCloudCreateRequestBody } from '@/validation-schemas/public-cloud';
 
 export default async function createOp({ session, body }: { session: Session; body: PublicCloudCreateRequestBody }) {
   const { user, permissions, ministries } = session;
