+++ app/app/api/public-cloud/products/_operations/update.ts
@@ -4,10 +4,10 @@ import { z, TypeOf, ZodType } from 'zod';
 import { BadRequestResponse, OkResponse, UnauthorizedResponse } from '@/core/responses';
 import { getPublicCloudProduct } from '@/queries/public-cloud-products';
 import editRequest from '@/request-actions/public-cloud/edit-request';
-import { PublicCloudEditRequestBody } from '@/schema';
 import { sendEditRequestEmails, sendExpenseAuthorityEmail } from '@/services/ches/public-cloud/email-handler';
 import { subscribeUsersToMautic } from '@/services/mautic';
 import { sendPublicCloudNatsMessage } from '@/services/nats';
+import { PublicCloudEditRequestBody } from '@/validation-schemas/public-cloud';
 import { putPathParamSchema } from '../[licencePlate]/schema';
 
 export default async function updateOp({
