+++ app/app/api/public-cloud/requests/search/route.ts
@@ -1,7 +1,7 @@
 import _isString from 'lodash-es/isString';
 import createApiHandler from '@/core/api-handler';
 import { OkResponse } from '@/core/responses';
-import { publicCloudRequestSearchBodySchema } from '@/schema';
+import { publicCloudRequestSearchBodySchema } from '@/validation-schemas/public-cloud';
 import searchOp from '../_operations/search';
 
 export const POST = createApiHandler({
