+++ app/constants/index.ts
@@ -1,4 +1,5 @@
 import { $Enums, Cluster, Ministry, Provider } from '@prisma/client';
+import { CpuQuota, MemoryQuota, StorageQuota } from '@/validation-schemas/private-cloud';
 
 export const clusters = Object.values(Cluster).filter((cluster) => cluster !== 'GOLDDR');
 
@@ -211,3 +212,7 @@ export const defaultProvisionedResourceValues = {
 };
 
 export const TEAM_SA_PREFIX = 'z_pltsvc-tsa-';
+
+export const cpuOptions = Object.values(CpuQuota.enum);
+export const memoryOptions = Object.values(MemoryQuota.enum);
+export const storageOptions = Object.values(StorageQuota.enum);
