+++ app/constants/regex.ts
@@ -1,0 +1,1 @@
+export const phoneNumberRegex = /[2-9]\d{2}[2-9](?!11)\d{6}$/;
